backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic-settings==2.7.0
pydantic[email]==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
authlib==1.3.2
httpx==0.28.1
google-api-python-client==2.160.0
google-auth-oauthlib==1.2.1
anthropic==0.43.0
cryptography==44.0.0
